Updated IWeighted to use double

+++ src/main/java/com/raduvoinea/utils/generic/dto/IWeighted.java
@@ -2,6 +2,6 @@ package com.raduvoinea.utils.generic.dto;
 
 public interface IWeighted {
 
-	int getWeight();
+	double getWeight();
 
 }
